+++ tests/upstream_alignment_test.py
@@ -37,6 +37,7 @@ class UpstreamAlignmentTest(TestCase):
         'mj-head-with-comment',
         'mj-image-with-empty-alt-attribute',
         'mj-image-with-href',
+        'mj-section-with-full-width',
         'mj-section-with-mj-class',
         'mj-section-with-background-url',
         'mj-font',