+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ------------------
 
 - add mjml-head-breakpoint, mjml-navbar, mjml-navbar-link, mjml-spacer, and mjml-wrapper components (Casey Holzer)
+- add support for custom components (Casey Holzer)
 
 
 0.8.0 (2022-08-25)